Refactor: 코드리팩도링
