+++ models/user.js
@@ -1,7 +1,16 @@
-const prismaClient = require('./prisma-client');
+const {PrismaClient} =require('@prisma/client')
+const prisma = new PrismaClient()
 
-function createUser() {}
+async function createUser(user_id,encryptedPassword,name) {
+   return await prisma.$queryRaw 
+  `INSERT INTO users (user_id,password,name) VALUES (${user_id},${encryptedPassword},${name})`;
+}
 
-function readUserByEmail(email) {}
+async function readuser(user_id){
+ const [id] = await prisma.$queryRaw
+    `SELECT * From users WHERE users.user_id = ${user_id}`;
+ return id;   
+}
 
-module.exports = { createUser, readUserByEmail };
+
+module.exports = { createUser, readuser };
